Train permutation tester
This script creates all possible classifier permutations on the "Train" page,
and then runs each test image against all of it's group's permutations to confirm the desired result.

This will help determine if https://github.ibm.com/Watson/developer-experience/issues/95 is a one-off thing or a systematic issue.

It may also be useful for cache warming and other updates to the demo.