Updates the README to describe common uses
The readme now explains how to update the sample images, includes pointers
on how to update the bundles. It also includes a feature to display
the custom classifier ID (by clicking on the classifier name) and
a hash handler to grab a custom classifier ID for use in the form.  It also
includes a new environment variable to always override the classifier ID and
only use a custom classifier.